Exporting sendgrid as separete module in gradle
