+++ backend/email-to-me/build.gradle
@@ -41,6 +41,7 @@ dependencies {
 
     compile (group: 'org.jetbrains.kotlin', name: 'kotlin-stdlib-jdk8', version: kotlin_version)
     compile (group: 'com.google.appengine', name: 'appengine-api-1.0-sdk', version: '1.9.60')
+    compile project(':sendgrid')
 
     compile group: 'javax.servlet', name: 'servlet-api', version: '2.5'
     compile group: 'com.google.code.gson', name: 'gson', version: '2.8.2'
