--- backend/src/main/webapp/googlec9a2ecc3bc85ebf9.html
@@ -1,1 +1,0 @@
-google-site-verification: googlec9a2ecc3bc85ebf9.html
